python_service/generate_document/requirements.txt
boto3
python-docx
aiohttp
